Funny shapes version and further optimizations. running at 60fps now, no dropped frames!
