+++ README.md
@@ -4,11 +4,18 @@ This project uses Parcel and TypeScript. Object creation and handling is done th
 
 ## TODO:
 * make it so that every shape processes a frame and then it is called at once by the manager??
+* babies have relationships init?? or no?
+* i can probably remove ColorManager
+* fix chain color (just take the averages)
+* fix chains connecting to multiple other shapes? make adultery a feature lol
+* optimization ideas: 
+  * path simplification
+  * layering for babies vs adults
+  * move complex shit out of the render loop
 * fix run loops! optimize code time complexity
 * clean up genital generation, fucking nasty
 * fix transition from baby to adult as the pivot point is messed up, they jump when transitioning
 * fix Relationship logic. lots of bugs are coming from the management of removing and adding partners, leaving unaddressed chains on screen or having shapes teleporting to NaN coords while seeking because the attractor is no longer in that relationship
-* fix colors and dot intermingling. dots will chain to others but not apply the relationship colors. just remove the grays and have every color be randomized to also make attraction calculations easier. 
 * optimizations. idk how to optimize this shit but it gets laggy super quick.
 
 * make drawing frames requested not always!
